+++ Tests/Functional/Traits/LanguageHelper.php
@@ -4,7 +4,10 @@ declare(strict_types=1);
 
 namespace OliverKlee\Seminars\Tests\Functional\Traits;
 
+use TYPO3\CMS\Core\Information\Typo3Version;
 use TYPO3\CMS\Core\Localization\LanguageService;
+use TYPO3\CMS\Core\Localization\LanguageServiceFactory;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
 
 /**
  * This trait provides methods useful for initializing languages.
@@ -19,7 +22,11 @@ trait LanguageHelper
     private function getLanguageService(): LanguageService
     {
         if (!$this->languageService instanceof LanguageService) {
-            $languageService = LanguageService::create('default');
+            if ((new Typo3Version())->getMajorVersion() >= 11) {
+                $languageService = GeneralUtility::makeInstance(LanguageServiceFactory::class)->create('default');
+            } else {
+                $languageService = LanguageService::create('default');
+            }
             $languageService->includeLLFile('EXT:seminars/Resources/Private/Language/locallang.xlf');
             $this->languageService = $languageService;
         }