+++ CHANGELOG.md
@@ -15,6 +15,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ### Removed
 
 ### Fixed
+- Avoid using the deprecated `LanguageService::create()` (#2712)
 - Avoid crash with Fluid 2.12.0 due to signature change (#3059)
 
 ## 5.5.1